// admcentro-web-main/src/interfaces/IpaymentType.ts
// Generated by https://quicktype.io

import { IClienyPayment } from './IclientPayments'

export interface IpaymentTypeResponse {
  code: number
  success: boolean
  message: string
  data: IpaymentType[]
}

export interface IpaymentType {
  id: number
  name: string
  createdAt?: string
  updatedAt?: string
}

// Generated by https://quicktype.io

export interface IStatisticsPayment {
  ok: boolean
  code: number
  status: string
  getTotalByPaymentType: GetTotalByPaymentType[]
  getTotalByMonth: GetTotalByMonth[]
  all: IClienyPayment[]
}

export interface GetTotalByMonth {
  month: string
  total: number
}

export interface GetTotalByPaymentType {
  PaymentTypeName: string
  total: number
}